Bug Admin corrigido ao deletar

+++ UniinfoAsp/UniinfoAsp/Controllers/LoginnController.cs
@@ -135,9 +135,18 @@ namespace UniinfoAsp.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Loginn loginn = db.Loginns.Find(id);
-            db.Loginns.Remove(loginn);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if(loginn.login.Equals("jw"))
+            {
+                ViewBag.Message = "Este login não pode ser apagado";
+
+            }
+            else
+            {
+                db.Loginns.Remove(loginn);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return View(loginn);
         }
 
         protected override void Dispose(bool disposing)
